Add custom map utility
Add custom map utility.
