+++ Assets/Scripts/Core/MyCustomMap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e1c788cc575d8454aa88fb4375fc9215
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
