+++ Assets/Scripts/LoadMap.cs
@@ -13,8 +13,6 @@ public class LoadMap : MonoBehaviour
 
     public int Players;
     public int PlayerId;
-
-    private Block[] map;
     public int size = 13;
 
     public void Awake()
@@ -109,76 +107,16 @@ public class LoadMap : MonoBehaviour
             }
         }
 
-        int half = size >> 1;
-        map = new Block[size * size];
-
-        var temp = new Vector2[]
-        {
-            new Vector2(1f, 1f),
-            new Vector2(size - 2, size - 2),
-            new Vector2(1f, size - 2),
-            new Vector2(size - 2, 1f),
-
-            new Vector2(half, half),
-
-            new Vector2(1f, half),
-            new Vector2(half, 1f),
-            new Vector2(size - 2, half),
-            new Vector2(half, size - 2),
-        };
-
-        for(int i = 0; i < size; i++)
-        {
-            for(int j = 0; j < size; j++)
-            {
-                int index = i * size + j;
-                map[index] = Block.Breakable;
-            }    
-        }
-
-        for(int i = 0; i < Players; i++)
-        {
-            var list = FindLocation(temp[i]);
-
-            for (int j = 0; j < list.Length; j++)
-            {
-                int x = Mathf.RoundToInt(list[j].x);
-                int y = Mathf.RoundToInt(list[j].y);
-                map[x * size + y] = Block.Born;
-            }
-        }
-
-        for(int i = 0; i < size; i++)
-        {
-            map[i] = Block.Wall;
-            map[(size - 1) * size + i] = Block.Wall;
-            map[i * size] = Block.Wall;
-            map[i * size + size - 1] = Block.Wall;
-        }
-
-        for (int i = 2; i < size; i += 2)
-        {
-            for (int j = 2; j < size; j += 2)
-            {
-                int index = i * size + j;
-                map[index] = Block.Wall;
-            }
-        }
+        MyCustomMap.CreateMap(size);
 
         for(int j = 0; j < size; j++)
         {
             for(int k = 0; k < size; k++)
             {
-                int index = j * size + k;
-                AddBlock(map[index], j, k);
+                Block block = MyCustomMap.GetBlock(j, k);
+                AddBlock(block, j, k);
             }
         }
-    }
-
-    public bool CanWalk(int j, int k)
-    {
-        int index = size * j + k;
-        return map[index] != Block.Wall;
     }
 
     private void AddBlock(Block block, int i, int j)
